Fixed files
removed fiery_srike and ki_bolt action files, they are properly called fiery_fist and ki_blast
